Move github-related functions to separate submodule for better structure
